Renamed vite folder to original name
