--- frontend-vite/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-.DS_Store
-dist
-dist-ssr
-*.local
