--- frontend-vite/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "frontend-vite",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "socket.io-client": "^4.4.1",
-    "socketio": "^1.0.0",
-    "vue": "^3.2.25",
-    "vue-3-socket.io": "^1.0.4",
-    "vue-router": "^4.0.12"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-vue": "^2.0.0",
-    "vite": "^2.7.2"
-  }
-}
