--- frontend-vite/src/App.vue
@@ -1,33 +1,0 @@
-<template>
-  <div id="nav">
-    <router-link to="/about">About</router-link> |
-    <router-link to="/list">List</router-link>
-  </div>
-  <router-view />
-</template>
-
-<style>
-html {
-  scroll-behavior: smooth;
-}
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-}
-
-#nav {
-  padding: 30px;
-}
-
-#nav a {
-  font-weight: bold;
-  color: #2c3e50;
-}
-
-#nav a.router-link-exact-active {
-  color: #42b983;
-}
-</style>
