+++ frontend/src/main.js
@@ -1,13 +1,18 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-import router from './router'
-import VueSocketIO from 'vue-3-socket.io'
-import io from 'socket.io-client'
-import './registerServiceWorker'
+import { createApp } from "vue";
+import App from "./App.vue";
+import router from "./router";
+import VueSocketIO from "vue-3-socket.io";
+import io from "socket.io-client";
 
 const socketio = new VueSocketIO({
-    debug: false,
-    connection: io('http://localhost:3000'),
-  });
+  debug: false,
+  connection: io("http://localhost:3000"),
+});
 
-createApp(App).use(router).use(socketio).mount('#app')
+const app = createApp(App);
+
+app.use(router);
+
+app.use(socketio);
+
+app.mount("#app");
