--- frontend/src/views/Pucko.vue
@@ -1,5 +1,0 @@
-<template>
-  <div class="about">
-    <h1>This is a pucko</h1>
-  </div>
-</template>
