--- frontend/vue.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-    devServer: {
-        disableHostCheck: true
-    }
-}
